making jspool gc-able
